new lib/languages.py, filter languages in filter_langlinks.py
