+++ tests/run.sh
@@ -1,4 +1,17 @@
 #!/bin/bash
+
+OUT=$(python3 -c'from lib.languages import Languages; print(len(Languages.get_languages()))')
+if [[ "$OUT" != "39" ]]; then
+    echo 'expected 39'
+    exit 1
+fi
+
+OUT=$(LANGUAGES=de,fr,it,en python3 -c'from lib.languages import Languages; print(len(Languages.get_languages()))')
+if [[ "$OUT" != "4" ]]; then
+    echo 'expected 4'
+    exit 1
+fi
+
 cat tests/filter_pagelinks.test1.txt | bin/filter_pagelinks.py > out.txt
 diff --brief out.txt tests/filter_pagelinks.test1expected.txt || exit 1
